requirements.txt
spacy>=3.5
medspacy
pandas
pyyaml
